스킬 슬롯 작업중
스킬 슬롯 작업중

+++ Project/Assets/Main/Scripts/FrameWork/Controller/Managers/MonsterManager.cs
@@ -112,7 +112,7 @@ namespace GameBerry.Managers
             }
         }
         //------------------------------------------------------------------------------------
-        public void OnDamage(double range, int damage, Vector3 playerpos)
+        public void OnDamage(float range, int damage, Vector3 playerpos)
         {
             var node = m_spawnedMonster_Linked.First;
             while (node != null)
